+++ src/Linq.Expression.Optimizer/AssemblyInfo.fs
@@ -5,13 +5,13 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("Linq.Expression.Optimizer")>]
 [<assembly: AssemblyProductAttribute("Linq.Expression.Optimizer")>]
 [<assembly: AssemblyDescriptionAttribute("Lightweight optimizer of System.Linq.Expression expressions. Just basic boolean algebra and reductions, constant and tuple/anonymous type eliminations.")>]
-[<assembly: AssemblyVersionAttribute("1.0.31")>]
-[<assembly: AssemblyFileVersionAttribute("1.0.31")>]
+[<assembly: AssemblyVersionAttribute("1.0.32")>]
+[<assembly: AssemblyFileVersionAttribute("1.0.32")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Linq.Expression.Optimizer"
     let [<Literal>] AssemblyProduct = "Linq.Expression.Optimizer"
     let [<Literal>] AssemblyDescription = "Lightweight optimizer of System.Linq.Expression expressions. Just basic boolean algebra and reductions, constant and tuple/anonymous type eliminations."
-    let [<Literal>] AssemblyVersion = "1.0.31"
-    let [<Literal>] AssemblyFileVersion = "1.0.31"
+    let [<Literal>] AssemblyVersion = "1.0.32"
+    let [<Literal>] AssemblyFileVersion = "1.0.32"